props: Refactor options
Boolean options are internally stored in an integer.
Initially the mapping from the public API's CERTIFIER_OPT_*
enum keys to option bits was a distance from a 'first' option
key. This change reworks the mapping to use an explicit lookup
table so boolean options don't have to be in a magical, contiguous
range in the public API, which is a maintenance hazard.

 * Add utility macros for [compile time constant] array length and
   integer<->pointer (up to 32 bits) conversions.
 * Look up option bits by public key, warn if not mapped.

Signed-off-by: Jesse Crews <jesse_crews@comcast.com>